==> Scripts/s02_dataset_review.Rmd <==
---
title: "Source data review"
author: "Helena Edlund"
date: "2109-01-03"
output: html_document
---

```{r setup, include=FALSE}
###################################################
# s02_dataset_review.Rmd
# Description: Review of rawdata, identification of a priori outliers etc
# Dependencies: s01_dataset_preparation.R / s01.RData
###################################################

# Settings to knit in top directory:
# Everything after this chunk works with paths relative top level
library(rprojroot)
knitr::opts_knit$set(root.dir=find_root(has_file("OpenProject.Rproj"))) 
knitr::opts_chunk$set(echo=F)

# Note: R markdown opens a new R session, your global environment is not available.
```


Load the dataset as prepared by "s01_dataset_preparation.R":
```{r, echo=T, message=F}
# -----------------------------------------------
# Prepare environment
# -----------------------------------------------
source(file = file.path("./Scripts","Setup","setup01_rEnvironment.R"))
load(file = file.path("./Scripts","s01.RData"))
```

# Structure and dimentions
The first 10 rows of the data: 
```{r, echo=T}
head(rawdata, 10)
```

The dimentions of the data:
```{r, echo=T}
dim(rawdata)
```

Included columns: 
```{r, echo=T}
names(rawdata)
```

Are there any duplicated rows?
```{r}
check_message(logical = any(duplicated(rawdata)))
```


# Cross-checks of project setup files and dataset

Are all columns in the dataset defined in "setup03_variables.R"?
```{r}
check_columns(names(rawdata), columns[['all']])
```

Are any columns defined in "setup03_variables.R" and not included in the dataset?
```{r}
check_columns(columns[['all']], names(rawdata))
```

Which studies are included and are they defined in "setup03_variables.R": 
```{r}
unique(rawdata$STUDYID)

check_columns(unique(rawdata$STUDYID), as.integer(studies))
```

Which dose groups are included and are they defined in "setup03_variables.R": 
```{r}
unique(rawdata$DOSE)  

check_columns(unique(rawdata$DOSE), as.integer(doses))
```

## Is the data sorted correctly?
Are the NMSEQSID defined in an increasing order?
```{r} 
ids      <- unique(rawdata$NMSEQSID)
id_shift <- c(ids[-1], NA)
id_diff  <- na.omit(ids-id_shift)

check_message(logical = all(id_diff < 0))
```

Is TIME increasing for each subject?
```{r} 
tmp <- 
  rawdata %>%
  group_by(NMSEQSID) %>%
  mutate(TIMESHIFT = c(TIME[-1], NA)) %>%
  mutate(TIMEDIFF = (TIME-TIMESHIFT)) %>%
  filter(!is.na(TIMEDIFF)) %>% 
  summarize(TIMEINCR = all(TIMEDIFF < 0)) %>% # summary for each patient
  summarize(output = unique(TIMEINCR))        # summary for entire dataset

check_message(logical = tmp$output)
```

Is TAPD increasing for each dose event?
```{r} 
# Explicitly add dose times to the dataset (expand addl and ii)
doses <- rawdata %>% 
  filter(!is.na(AMT)) %>% 
  select(NMSEQSID, TIME, AMT, ADDL, II)  

for(i in 1:nrow(doses)){
  dosetime <- seq(from = doses$TIME[i], 
                  by   = doses$II[i], 
                  length.out = (doses$ADDL[i]+1))
  
  dat <- data.frame(NMSEQSID = rep(doses$NMSEQSID[i], length(dosetime)), 
                    TIME = dosetime, 
                    AMT = rep(doses$AMT[i], length(dosetime)))
  if(i == 1){
    dat_out <- dat
  } else {
    dat_out <- bind_rows(dat_out, dat)
  }
} 

# Merge 
tmp <- rawdata %>% 
  full_join(dat_out) %>% 
  arrange(NMSEQSID, TIME)

rm(dat, dat_out, dosetime, doses)


# Add a flag identifier for each dosing event (can be checked irrespectively of subject id)
# 1. Check if there are any rows before the first dose. 
na_index <- which(!is.na(tmp$AMT))[1]-1
# 2. Create flag
tmp <- tmp %>% 
  mutate(AMTFLAG = ifelse(!is.na(AMT), c(1:n()), NA), 
         AMTFLAG = ifelse(na_index==0, 
                          na.locf(AMTFLAG), 
                          # if rows before first dose, fill those with NA and do locf for the rest
                          c(rep(NA, na_index), na.locf(AMTFLAG)))
  )

# Summarize similar to above
tmp <- tmp %>%
  group_by(AMTFLAG) %>%
  mutate(TAPDSHIFT = c(TAPD[-1], NA)) %>%
  mutate(TAPDDIFF = (TAPD-TAPDSHIFT)) %>%
  filter(!is.na(TAPDDIFF)) %>% 
  summarize(TAPDINCR = all(TAPDDIFF < 0)) %>% 
  summarize(output = unique(TAPDINCR))

check_message(logical = tmp$output)
```



## Subjects

The total number of subjects included is:
```{r}  
nTotal <- length(unique(rawdata$NMSEQSID))
nTotal
```

The total number of subjects in each study:
```{r}  
rawdata %>% 
  group_by(STUDYID) %>% 
  summarize(N=length(unique(NMSEQSID)))
```

Number of PK occasions and maximum follow up time per subject:
```{r}  
rawdata %>% 
  group_by(STUDYID, NMSEQSID) %>% 
  summarize(nOcc=ifelse(all(is.na(OCC)), 
                        0, max(unique(OCC[is.na(AMT)]), na.rm=T)),
            maxTIME=max(TIME, na.rm=T)) %>% 
  summarize(min_occ_per_id = min(nOcc), 
            max_occ_per_id = max(nOcc), 
            min_follow_up_per_id = min(maxTIME), 
            max_follow_up_per_id = max(maxTIME))
```



## Regimens
Does any subject have any (unexpected) change in dosing freqency?
```{r} 
# Don't forget to exclude cohorts where you expect a change (if any)
tmp <- rawdata %>% 
  group_by(STUDYID, NMSEQSID) %>% 
  filter(II != 0) %>% 
  summarize(II = collapse_unique(II)) %>% 
  summarize(duplicates = unique(duplicated(NMSEQSID))) 

check_message(logical = any(tmp$duplicates))
``` 

Does any subject have any (unexpected) change in dose?
```{r} 
# Don't forget to exclude cohorts where you expect a change (if any)
tmp <- rawdata %>% 
  group_by(STUDYID, NMSEQSID) %>% 
  filter(!is.na(AMT)) %>% 
  summarize(AMT = collapse_unique(AMT)) %>% 
  summarize(duplicates=unique(duplicated(NMSEQSID))) 

check_message(logical = any(tmp$duplicates))
```


## Column summaries (not including covariates)
Summary of numeric columns:
```{r}
rawdata %>% 
  select(columns[['numeric']]) %>% 
  summary()
```

Summary of numeric columns by study:
```{r}
for(i in studies){
  print(paste("study", i,":"))
  print(
    rawdata %>% 
      filter(STUDYID==i) %>% select(columns[['numeric']]) %>% 
      summary()
    )
}
```

Unique values of categorical columns:
```{r}
for(i in columns[['factors']]){
  print(rawdata %>% select(i) %>% unique())
}
```

Unique values of categorical columns by study:
```{r}
for(i in columns[['factors']]){
print(rawdata %>% group_by(STUDYID) %>%  
        select(c("STUDYID", i)) %>% unique())
}
```


## Cross-check of AMT, EVID, DV, MDV, BQL and CMT:
Is EVID==0 when AMT is missing?
```{r}
test <- rawdata %>% 
  filter(is.na(AMT)) %>% 
  summarise(evid = collapse_unique(EVID))

check_message(logical = test$evid==1)
```

Is EVID==1 when AMT is *not* missing?
```{r}
test <- rawdata %>% 
  filter(!is.na(AMT)) %>% 
  summarise(evid = collapse_unique(EVID))
            
check_message(logical = test$evid==1)
```

Is EVID=0 for observations?
```{r}
test <- rawdata %>% 
  filter(!is.na(DV)) %>% 
  summarize(evid = collapse_unique(EVID))

check_message(logical = test$evid==0)
```

Is BLQ=0 for non-missing observations?
```{r}
test <- rawdata %>% 
  filter(!is.na(DV)) %>% 
  summarize(blq = collapse_unique(BLQ))
check_message(logical = test$blq==0)
```

Are any quantified concentrations reported $<$LLOQ?
```{r}
test <- rawdata %>% 
  filter(!is.na(DV)) %>% 
  summarize(blq_dv = collapse_unique(DV < LLOQ)) 

check_message(logical = test$blq_dv=="FALSE")
```

Are all TAPD=0 dosing events (evid=1)?
```{r}
test <- rawdata %>% 
  filter(!is.na(TAPD) & TAPD==0) %>% 
  summarize(evid = collapse_unique(EVID))

check_message(logical = test$evid==1)
```

Are the compartments set correctly?
```{r}
rawdata %>% 
  group_by(EVID) %>% 
  summarize(cmt = collapse_unique(CMT)) 
```

Cross-check columns for rows with missing DV:
```{r}
rawdata %>% 
  subset(is.na(DV) & is.na(AMT)) %>% 
  summarize(dv = collapse_unique(DV), 
            evid = collapse_unique(EVID), 
            mdv = collapse_unique(MDV), 
            blq = collapse_unique(BLQ))
```




## Cross-check of C and COMMENTS
Is there any C="C" which does *not* have an exclusion comment?
```{r}
test <- rawdata %>% 
  filter(C=="C") %>% 
  summarise(comment = any(COMMENT %in% c("", " ") | is.na(COMMENT)))

check_message(logical = test$comment)
```



## Covariate summary and cross-check
### Baseline data
```{r}
baseline_rawdata <- rawdata %>% 
  subset(!duplicated(NMSEQSID))
```

Summary of continous covariates:
```{r}
summary(baseline_rawdata[, columns[['base_cont_cov']]])
``` 
Unique values of categorical covariates
```{r}
apply(baseline_rawdata[, columns[['base_cat_cov']]], 2, unique)
```

Is there only one value for each subject? (should return TRUE for all subjects)
```{r}
# Each individual should only have one value
per_id <- function(x){
  return(length(unique(x))==1) 
}
summary(
  rawdata %>% 
  group_by(NMSEQSID) %>% 
  summarize_at(c(columns[['base_cont_cov']],columns[['base_cat_cov']]), 
               funs(values=per_id))
  )
```

#### Cross-check of covariate categories to ensure that the coding seems correct:
Does males have higher body weight?
```{r}
baseline_rawdata %>% 
  filter(BWT !=-99 | !is.na(BWT) ) %>% 
  group_by(SEXM) %>% 
  summarise(min = min(BWT),
            median = median(BWT),
            max = max(BWT))
```
