==> Scripts/s06_nm_datasets.Rmd <==
---
title: "NONMEM dataset preparation"
author: "Helena Edlund"
date: "2019-01-07"
output: 
  html_document:
params:
  print_csv: FALSE
---

```{r setup, include=F}
###################################################
# s06_nm_datasets.Rmd
# Description: Dataset preparation for NONMEM
# Dependencies: s01_dataset_preparation.R / s01.RData 
###################################################

# Settings to knit in top directory:
# Everything after this chunk works with paths relative top level
library(rprojroot)
knitr::opts_knit$set(root.dir=find_root(has_file("OpenProject.Rproj"))) 
knitr::opts_chunk$set(echo=F)

# Note: R markdown opens a new R session, your global environment is not available.
```

This script uses the data.frame "rawdata". That is, all rows with C=="C" are included (as required by regulatory agencies). These rows are excluded in the NONMEM scripts instead. 

```{r environmentPrep, message=F}
# -----------------------------------------------
# Prepare environment
# -----------------------------------------------
source(file = file.path("./Scripts","Setup","setup01_rEnvironment.R"))
load(file = file.path("./Scripts","s01.RData"))

```

Are the datasets being outputted to file?
```{r, echo=F}
params$print_csv
```

# Selection of columns and re-naming for NONMEM 
```{r}
# List numeric columns to include for nonmem dataset
nm_columns <- c("C","NMSEQSID","TIME","TAPD","AMT","DV","MDV","EVID",
                "ADDL","II","CMT","BLQ","OCC","STUDYID","DOSE","DAY",
                "AGE","BCRCL","BWT","SEXM","RACE")

# use rawdata: contains all "C" and numeric versions of columns
nm_data <- rawdata %>% select(nm_columns)

# rename NMSEQSID to ID 
nm_data <- nm_data %>% 
  rename(ID = NMSEQSID)

## Add a RATE column with values -2 at dosing records 
# for estimation of zero order input (for models with zero order input)
nm_data <- nm_data %>% 
  mutate(RATE = ifelse(!is.na(AMT), -2, NA))

```

# Imputation of missing covariates

```{r}
## 1. Categorical covariates: model as separate groups and 
# potentially merge with other (in nonmem script)
nm_data <- nm_data %>% 
  mutate(RACE = ifelse(is.na(RACE), -99, RACE))


## 2. Continous covariates: impute based on median
impute_baseline <- 
  nm_data %>% 
  filter(!duplicated(ID)) %>% 
  summarize_at(columns[["base_cont_cov"]],
               .funs = "median", na.rm=T)

nm_data <- nm_data %>% 
  mutate(
    # Age
    AGE = ifelse(is.na(AGE), impute_baseline$AGE, AGE))
    
```


# Output and naming of dataset
The dataset names are saved in variables that are used to update the nonmem control-files later on.

## No. 1: Original with imputed covariates
```{r}
# Dataset name and path to location
nm_data_filename <- paste0(dv_name,"_nm_", delivery_date,".csv")
out_file         <- file.path(directories[["derived_data_dir"]], nm_data_filename)

# Write dataset 
if(params$print_csv){ 
  write.csv(nm_data, file=out_file, 
            row.names=F, quote=F, na = ".") 
}
```

# List of prepared datasets
```{r, echo=F}
# Save dataframe with number of datasets and modifications done. 
nm_datasets <- data.frame(
  filename = nm_data_filename, 
  Description = c("Dataset prepared for NONMEM use: (i) only numerical columns selected, (ii) missing covariates imputed, (iii) NMSEQSID renamed to ID, (iv) RATE column added to estimate duration of zero-order absorption")
  )

kable(nm_datasets)
```


```{r, echo=F}
# Write dataset 
if(params$print_csv){
  write.csv(nm_datasets, 
            file=file.path(directories[["derived_data_dir"]], "prepared_datasets.csv"), 
            row.names=F, quote=T)
}
```


```{r, include=F}
# Save environment to be used in developement scripts 
rm("params") # remove rmarkdown params or it will cause issues in the next script

# this will generate in error but all the output is generated as should.
save.image(file = file.path(directories[["scripts_dir"]],"s06.RData"))
```
